+++ ricaun.Nuke/Components/ICompile.cs
@@ -3,13 +3,13 @@ using ricaun.Nuke.Extensions;
 
 namespace ricaun.Nuke.Components
 {
-    public interface ICompile : IClean, IHazSolution, INukeBuild
+    public interface ICompile : IClean, IHazMainProject, IHazSolution, INukeBuild
     {
         Target Compile => _ => _
             .DependsOn(Clean)
             .Executes(() =>
             {
-                Solution.BuildMainProject();
+                Solution.BuildProject(MainProject);
             });
     }
 }