+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.0.10] / 2021-12-17
+- Change Base Release `.nupkg` or full folder
+- Fix HazSolutionExtension
+- Add GetMainProject on IHazMainProject 
+- private GetMainProject
+- Add HazSolutionExtension
+- Add HazMainProjectExtension
+- Add IHazMainProject
+- Rename Example Folder Project
+
 ## [0.0.9] / 2021-12-17
 - Clear Build
 - Test Native Example Compile
@@ -66,6 +76,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: https://github.com/ricaun-io/ricaun.Nuke/compare/1.0.0...HEAD
+[0.0.10]: https://github.com/ricaun-io/ricaun.Nuke/compare/0.0.9...0.0.10
+[0.0.9]: https://github.com/ricaun-io/ricaun.Nuke/compare/0.0.8...0.0.9
 [0.0.8]: https://github.com/ricaun-io/ricaun.Nuke/compare/0.0.7...0.0.8
 [0.0.7]: https://github.com/ricaun-io/ricaun.Nuke/compare/0.0.6...0.0.7
 [0.0.6]: https://github.com/ricaun-io/ricaun.Nuke/compare/0.0.5...0.0.6